Kept Flip's old IMWPaths module but created a new IMW::Paths module which is intended to also provide mixin functionality.
Right now it implements a basic path finding method for Datsets and DataSources with regard to a particular step in the IMW workflow.  It doesn't interface with the IMW::Config module so it doesn't actually know about locally configured paths to IMW directories and so is quite stupid in that sense but at least it seems to work.